Added save function to external database
User will be prompted to choose a database file for the data to be saved in - if they choose a different file type the system will tell them to pick a database file.
If the database is set up wrong (not got the right table), the system should tell them to make sure it's set up right. Check this in testing